Feat: broke up code into 2 classes'
