+++ spec/spec_helper.rb
@@ -1,4 +1,5 @@
 require './lib/board'
+require './lib/game_functions'
 require './lib/connect_four'
